fix uint64

+++ Tether.cs
@@ -110,7 +110,7 @@ namespace Palantir
         {
             try
             {
-                await (await Program.Client.GetChannelAsync(Convert.ToUInt32(PalantirEndpoint.ChannelID))).SendMessageAsync("The observed message couldn't be found. \nSet a channel using `>observe #channel`!");
+                await (await Program.Client.GetChannelAsync(Convert.ToUInt64(PalantirEndpoint.ChannelID))).SendMessageAsync("The observed message couldn't be found. \nSet a channel using `>observe #channel`!");
             }
             catch(Exception e)
             {
